requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.1.0
Django==6.0.1
django-cloudinary-storage==0.3.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.7
gunicorn==24.1.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyYAML==6.0.3
qrcode==8.2
render==1.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
